# crates/public-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "canopy public-server",
    "description": "Internet-facing API for the canopy fleet. Device-authenticated endpoints require an mTLS client certificate.\n\nRequest bodies may be sent compressed, and this is recommended for large payloads: the server transparently decodes `Content-Encoding: gzip`, `br`, `deflate`, or `zstd`. The accepted encodings are also listed structurally in the `x-request-compression` extension.",
    "contact": {
      "name": "BES Developers",
      "email": "contact@bes.au"
    },
    "license": {
      "name": "GPL-3.0-or-later"
    },
    "version": ""
  },
  "paths": {
...
    "/servers": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List publicly-listed central applications.",
        "description": "Returns every central server that has both a public display name and a\nreachable host configured, ordered by environment tier (production\nfirst, then clone, demo, test, dev) and then by name. Used by clients\nto let a user pick which server to connect to.",
        "operationId": "list_servers",
        "responses": {
          "200": {
            "description": "Publicly-listed central applications, ordered by rank then name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicServer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/versions/{version}": {
      "post": {
        "tags": [
          "versions"
        ],
        "summary": "Publish a version with its changelog.",
        "description": "Requires a device certificate with the releaser role (or admin). The\npath parameter is the exact version being published (e.g. `2.10.5`);\nthe request body is the changelog for that version, as up to 1 MiB of\nmarkdown text.\n\nIf the version already exists in the draft state — for example\nbecause an artifact was registered against it before its changelog\nwas written — the draft is published in place, with this changelog\nreplacing whatever it had before. Otherwise a new version is created\ndirectly in the published state. Publishing a version that already\nexists and is not a draft (already published, or yanked) fails.\n\nReturns the resulting version record.",
        "operationId": "create_version",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Changelog markdown text, up to 1 MiB.",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "releaser-device": []
          }
        ]
      },
      "delete": {
        "tags": [
          "versions"
        ],
        "summary": "Yank a version.",
        "description": "Requires a device certificate with the admin role. Marks the given\nexact version as yanked, hiding it from listings, update checks, and\nartifact lookups without deleting its history.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Version marked yanked."
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-device": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetailsSchema": {
        "type": "object",
        "description": "Standard error response body, returned for every non-2xx response.\n\nThis follows the RFC 7807 \"Problem Details\" shape: a stable machine-readable\n`type`, a short `title`, the repeated HTTP `status` code, and an optional\n`detail` string with specifics of this particular occurrence.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable explanation specific to this occurrence of the\nproblem, if any extra detail is available."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code of the response, repeated here for convenience.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type. Does not vary\nbetween occurrences of the same `type`."
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference identifying the problem type. Stable across\noccurrences of the same error, so callers can match on it.",
            "example": "/errors/resource-not-found"
          }
        }
      },
...
      "PublicServer": {
        "type": "object",
        "description": "A publicly-listed central server that a client can connect to.",
        "required": [
          "name",
          "host"
        ],
        "properties": {
          "host": {
            "$ref": "#/components/schemas/UrlField",
            "description": "The server's reachable base URL."
          },
          "name": {
            "type": "string",
            "description": "Public-facing display name of the server."
          },
          "rank": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerRank",
                "description": "The server's environment tier (production, clone, demo, test, or\ndev), if set. Used to order the listing and to let clients label\nnon-production entries."
              }
            ]
          }
        }
      },
...
      "ServerRank": {
        "type": "string",
        "description": "The environment tier of a server, from `production` down to `dev`.",
        "enum": [
          "production",
          "clone",
          "demo",
          "test",
          "dev"
        ]
      },
...
      "UrlField": {
        "type": "string",
        "format": "uri",
        "description": "A URL, given as a plain string. Any trailing slash is stripped when the\nvalue is returned."
      },
...
      "Version": {
        "type": "object",
        "description": "A release version of the monitored software, with its publication\nstatus and changelog.",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "major",
          "minor",
          "patch",
          "status",
          "changelog"
        ],
        "properties": {
          "changelog": {
            "type": "string",
            "description": "Changelog text for this version, as Markdown."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the version record was created."
          },
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The releaser device that published this version, if it was published\nby a device rather than created by an operator."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the version."
          },
          "major": {
            "type": "integer",
            "format": "int32",
            "description": "Major version number."
          },
          "minor": {
            "type": "integer",
            "format": "int32",
            "description": "Minor version number."
          },
          "patch": {
            "type": "integer",
            "format": "int32",
            "description": "Patch version number."
          },
          "status": {
            "$ref": "#/components/schemas/VersionStatus",
            "description": "Publication status: `draft`, `published`, or `yanked`."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the version record was last changed."
          }
        }
      },
      "VersionStatus": {
        "type": "string",
        "description": "Publication status of a release version.",
        "enum": [
          "draft",
          "published",
          "yanked"
        ]
      },
...
    },
    "securitySchemes": {
      "admin-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `admin` role."
      },
      "backup-restore-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `backup-restore` role (or `admin`)."
      },
      "releaser-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `releaser` role (or `admin`)."
      },
      "server-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `server` role (or `admin`)."
      }
    }
  },
...
}